Pushed downwards the stack scheduling of dispatched actions to channel's subscribe function, so store-like objects used by runSaga benefit the same scheduling fixes done previously in redux middleware
